+++ src/main/java/kr/co/souso/souso/global/utils/convert/Constant.java
@@ -1,4 +1,4 @@
-package kr.co.souso.souso.global.utils;
+package kr.co.souso.souso.global.utils.convert;
 
 public interface Constant {
